completato sviluppo twig-form e gestione campi unique

+++ public/js/main.js
@@ -1,4 +1,5 @@
 const statistics = document.getElementById('statistics');
+const elements = document.getElementById('elements');
 
 if(statistics){
     statistics.addEventListener('click', e => {
@@ -7,6 +8,18 @@ if(statistics){
                 const id = e.target.getAttribute('data-id');
                 console.log("statistic id: ", id);
                 fetch(`/statistica/rete/delete/${id}`, {
+                    method: 'DELETE'
+                }).then(res => window.location.reload());
+            }
+        }
+    } ); 
+} else if (netelements){
+    netelements.addEventListener('click', e => {
+        if(e.target.className === 'btn btn-danger'){
+           if(confirm('Are you sure to delete the Network Element?')){
+                const id = e.target.getAttribute('data-id');
+                console.log("Network Element id: ", id);
+                fetch(`/network/element/delete/${id}`, {
                     method: 'DELETE'
                 }).then(res => window.location.reload());
             }
